requirements.txt
# --- Audio Processing ---
librosa
pydub
soundfile
numpy

# --- Data Handling ---
pandas

# --- Whisper ASR ---
openai-whisper
ffmpeg-python

# --- Segmentation ---
nltk
sentence-transformers
scikit-learn
yake

# --- Summarization ---
transformers
torch

# --- Evaluation ---
jiwer

# --- Optional helpful utilities ---
tqdm
